TaskController routes & User slight modifications on routes

+++ src/main/kotlin/com/taskcode/service/TaskService.kt
@@ -2,13 +2,16 @@ package com.taskcode.service
 
 import com.taskcode.model.Task
 import com.taskcode.repository.TaskRepository
+import jakarta.persistence.EntityNotFoundException
 import org.springframework.stereotype.Service
 
 @Service
 class TaskService(private val taskRepository: TaskRepository) {
 
     fun getTaskById(id: Long) : Task? {
-        return taskRepository.findById(id).orElse(null)
+        return taskRepository.findById(id).orElseThrow {
+            EntityNotFoundException("Task Id '$id' not found")
+        }
     }
 
     fun saveTask(task: Task): Task {
@@ -16,6 +19,9 @@ class TaskService(private val taskRepository: TaskRepository) {
     }
 
     fun deleteTask(id: Long) {
-        return taskRepository.deleteById(id)
+        if (!taskRepository.existsById(id)) {
+            throw EntityNotFoundException("Task Id '$id' not found")
+        }
+        taskRepository.deleteById(id)
     }
 }
